Developing int64 display data

+++ omodscan/quintvalidator.h
@@ -12,13 +12,13 @@ class QUIntValidator : public QValidator
     Q_OBJECT
 public:
     explicit QUIntValidator(QObject *parent = nullptr);
-    QUIntValidator(uint bottom, uint top, QObject *parent = nullptr);
+    QUIntValidator(quint64 bottom, quint64 top, QObject *parent = nullptr);
 
     State validate(QString &, int &) const override;
 
 private:
-    uint _bottom;
-    uint _top;
+    quint64 _bottom;
+    quint64 _top;
 };
 
 #endif // QUINTVALIDATOR_H
